+++ extensions/folder-search/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Folder Search Changelog
 
+## [New Features] - 2022-09-14
+
+- **added** Plugins support (see README for details)
+- **added** Create Quicklink action
+- **added** Open With... action
+- **added** Show Info in Finder action
+
 ## [New Feature] - 2022-09-09
 
 - **added** Ability to 'Pin' folders